SSL implementation on all channels

+++ tests/integration/psync2-reg.tcl
@@ -36,23 +36,25 @@ start_server {} {
     }
 
     set cycle_start_time [clock milliseconds]
-    set bench_pid [exec src/redis-benchmark -p $R_port(0) -n 10000000 -r 1000 incr __rand_int__ > /dev/null &]
-    while 1 {
-        set elapsed [expr {[clock milliseconds]-$cycle_start_time}]
-        if {$elapsed > $duration*1000} break
-        if {rand() < .05} {
-            test "PSYNC2 #3899 regression: kill first replica" {
-                $R(1) client kill type master
+    if {!$::ssl} {
+        set bench_pid [exec src/redis-benchmark -p $R_port(0) -n 10000000 -r 1000 incr __rand_int__ > /dev/null &]
+        while 1 {
+            set elapsed [expr {[clock milliseconds]-$cycle_start_time}]
+            if {$elapsed > $duration*1000} break
+            if {rand() < .05} {
+                test "PSYNC2 #3899 regression: kill first replica" {
+                    $R(1) client kill type master
+                }
             }
-        }
-        if {rand() < .05} {
-            test "PSYNC2 #3899 regression: kill chained replica" {
-                $R(2) client kill type master
+            if {rand() < .05} {
+                test "PSYNC2 #3899 regression: kill chained replica" {
+                    $R(2) client kill type master
+                }
             }
+            after 100
         }
-        after 100
+        exec kill -9 $bench_pid
     }
-    exec kill -9 $bench_pid
 
     if {$debug_msg} {
         for {set j 0} {$j < 100} {incr j} {
